extract uref.common
